% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\docType{data}
\name{cal}
\alias{cal}
\title{Calibration data}
\format{A data.frame of 150 observations containing columns:\cr
    - \code{sampleid}\tab unique identifier for each sample;\cr
    - \code{plotid}\tab unique identifier for each sample location;\cr
    - \code{date}\tab sampling date;\cr
    - \code{year}\tab sampling year;\cr
    - \code{species}\tab 6-digit abbreviation for each species;\cr
    - \code{fg}\tab abbreviation for functional groups;\cr
    - \code{type}\tab moss 'm' or lichen 'l' organism type;\cr
    - \code{ovendrymass}\tab oven-dry mass, in grams;\cr
    - \code{depth}\tab vertical depth, in centimeters;\cr
    - \code{area}\tab horizontal area, in square centimeters;\cr
    - \code{c}\tab total carbon content, as a percentage;\cr
    - \code{n}\tab total nitrogen content, as a percentage;\cr
    - \code{vol}\tab volume, in cubic centimeters;\cr
    - \code{dens}\tab density, in grams per cubic centimeter.\cr}
\usage{
data(cal)
}
\description{
Calibration data using 150 destructive samples, used to build a
    calibration model for non-destructive estimation of ground
    layer biomass, carbon and nitrogen content.  Samples were
    collected in 2012 and 2013 from Oregon and Alaska, USA.
}
\details{
Destructive samples were collected to parameterize a nondestructive
     estimation model for the US Forest Service's Forest Inventory
     and Analysis program.  The specific FIA protocol is the
     'Ground Layer Indicator' (FIA 2018), first implemented as part
     of the Interior Alaska Inventory beginning in 2014.  The
     calibration model (a negative exponential function) is fully
     parameterized and described in Smith et al. (2015).  The
     simplified functional groups in this dataset must be
     crosswalked to their expanded versions prior to estimation.
}
\references{
Smith, R.J., J.C. Benavides, S. Jovan, M. Amacher, and B. McCune.
     2015. A rapid method for landscape assessment of carbon
     storage and ecosystem function in moss and lichen ground
     layers. The Bryologist 118(1): 32–45.

FIA [Forest Inventory and Analysis Program]. 2018. Field
     Instructions for the Annual Inventory of Alaska 2018:
     Supplement for Interior Alaska. USDA Forest Service, Pacific
     Northwest Research Station, Anchorage, Alaska.
}
\keyword{datasets}
