% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass.R
\name{calc_biomass}
\alias{calc_biomass}
\title{Calculate biomass}
\usage{
calc_biomass(x, ...)
}
\arguments{
\item{x}{data.frame, minimally containing the columns
\code{c("pid","microquad","gf","cover","depth")}.  Each row is
an observation for each functional grp within a microplot,
nested within transects > subplots > plots.}

\item{...}{further arguments (currently ignored).}
}
\value{
A data.frame of same number of rows as `x`, with new columns:
   'mpid', 'type', 'predCN', 'predN', 'predC', 'mass', 'dens',
   'nvalue', 'cvalue', 'volume', 'covercm'.
}
\description{
Negative exponential model to calculate biomass while accounting
    for density variation with ground layer depth.
}
\details{
Returns estimates of biomass, organic carbon, total nitrogen, and
     C:N ratio for each functional group observed per microquad.
     Equations follow Smith et al. (2015).  Example usage is in
     Smith et al. (2017).
}
\examples{
data(est)
x <- calc_biomass(est)
head(x)

}
\references{
Smith, R.J., J.C. Benavides, S. Jovan, M. Amacher, and B. McCune.
     2015. A rapid method for landscape assessment of carbon
     storage and ecosystem function in moss and lichen ground
     layers. The Bryologist 118(1): 32–45.

Smith, R. J., S. Jovan, A. N. Gray, and B. McCune. 2017.
    Sensitivity of carbon stores in boreal forest moss mats -
    effects of vegetation, topography and climate. Plant and Soil
    421:31–42.
}
