% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dat.R
\name{check_dat}
\alias{check_dat}
\title{Check data}
\usage{
check_dat(x, ...)
}
\arguments{
\item{x}{data.frame containing the columns
\code{c("pid","microquad","gf","cover","depth")}.}

\item{...}{further arguments passed to other functions.}
}
\value{
Logical value.
}
\description{
Verifies proper data format before further processing.
}
\examples{
x <- matrix(NA, 5, 5)
dimnames(x)[[2]] <- c("pid","microquad","gf","cover","depth")
check_dat(x)
check_dat(data.frame(x))

}
