% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\docType{data}
\name{est}
\alias{est}
\title{Estimation data}
\format{A data.frame of 6256 observations of functional groups in
     microquads.}
\usage{
data(est)
}
\description{
Estimation data at 96 plots measured in 2014 from interior Alaska,
     USA.
}
\details{
Non-destructive measurements were collected for the US Forest
     Service's Forest Inventory and Analysis program.  The specific
     FIA protocol is the 'Ground Layer Indicator' (FIA 2018), first
     implemented as part of the Interior Alaska Inventory beginning
     in 2014.  Plot coordinates have been fuzzed.  See Smith et al.
     (2015) and Smith et al. (2017) for use in publications.
}
\references{
Smith, R.J., J.C. Benavides, S. Jovan, M. Amacher, and B. McCune.
     2015. A rapid method for landscape assessment of carbon
     storage and ecosystem function in moss and lichen ground
     layers. The Bryologist 118(1): 32–45.

Smith, R. J., S. Jovan, A. N. Gray, and B. McCune. 2017.
    Sensitivity of carbon stores in boreal forest moss mats -
    effects of vegetation, topography and climate. Plant and Soil
    421:31–42.

FIA [Forest Inventory and Analysis Program]. 2018. Field
     Instructions for the Annual Inventory of Alaska 2018:
     Supplement for Interior Alaska. USDA Forest Service, Pacific
     Northwest Research Station, Anchorage, Alaska.
}
\keyword{datasets}
