% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grlyr.R
\docType{package}
\name{grlyr}
\alias{grlyr}
\alias{grlyr-package}
\title{grlyr: A package for ground layer estimation.}
\description{
Estimates biomass, carbon and nitrogen of moss and lichen ground
     layers as part of the US Forest Service Forest Inventory and
     Analysis program.
}
\section{Datasets}{


     \code{\link{cal}} is calibration data from destructive
     sampling.
     \code{\link{est}} is estimation data from nondestructive
     sampling at sites where estimates are desired.
}

\section{Functions}{


     \code{\link{check_dat}} checks that estimation data are in
     proper format.\cr
     \code{\link{calc_biomass}} is the core function to calculate
     ground layer metrics (prominently including biomass).\cr
     \code{\link{summary_plot}} produces summaries at the
     plot-level, the expected level of summarization for USFS-FIA.
     \cr
     \code{\link{summary_fg}} produces summaries at the level of
     functional groups.\cr
     \code{\link{plot_map}} plots a map of geographic locations
     colored by some value.\cr
     \code{\link{plot_facet}} plots a scatterplot or histogram of
     values faceted by some grouping variable (such as functional
     groups).
}

\section{Bibliography}{


     FIA [Forest Inventory and Analysis Program]. 2018. Field
     Instructions for the Annual Inventory of Alaska 2018:
     Supplement for Interior Alaska. USDA Forest Service, Pacific
     Northwest Research Station, Anchorage, Alaska.

     Pattison, R., H.-E. Andersen, A. Gray, B. Schulz, R. J. Smith,
     and S. Jovan. 2018. Forests of the Tanana Valley State Forest
     and Tetlin National Wildlife Refuge, Alaska: results of the
     2014 pilot inventory. Page 80. Gen. Tech. Rep. PNW-GTR-967, US
     Department of Agriculture, Forest Service, Pacific Northwest
     Research Station, Portland, OR.

     Smith, R. J., J. C. Benavides, S. Jovan, M. Amacher, and B.
     McCune. 2015. A rapid method for landscape assessment of
     carbon storage and ecosystem function in moss and lichen
     ground layers. The Bryologist 118:32–45.

     Smith, R. J., S. Jovan, A. N. Gray, and B. McCune. 2017.
     Sensitivity of carbon stores in boreal forest moss mats -
     effects of vegetation, topography and climate. Plant and Soil
     421:31–42.

     Smith, R. J., S. Jovan, and B. McCune. 2014. Ubiquitous moss
     and lichen mats promote forest health. Pages 68–69 in E.
     Graham and T. Huette, editors. Forest Health Conditions in
     Alaska 2013. R10-PR-035. USDA Forest Service, Alaska Region,
     Anchorage, Alaska.
}

