% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch web app}
\usage{
launch_app()
}
\value{
Nothing, simply opens another window with the web app.
}
\description{
Launch the web app to calculate biomass while accounting
    for density variation with ground layer depth.
}
\details{
Web app allows an external user to upload a file of observations,
    then download an automatically generated file, without getting
    involved in any modeling or analysis.
}
\examples{
\dontrun{
    launch_app()
}

}
