% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facet.R
\name{plot_facet}
\alias{plot_facet}
\title{Facet plot}
\usage{
plot_facet(data, x, y, by, free = "none", type = "plot",
  qline = TRUE, ...)
}
\arguments{
\item{data}{data.frame in long format.}

\item{x, y}{name of variable in \code{data} for which values are
plotted.}

\item{by}{name of grouping variable in \code{data}.}

\item{free}{one of \code{c('none','x','y','xy')}, defining which
axes will be freely varying; otherwise axes will be identical
among all facets.}

\item{type}{one of \code{'plot'} or \code{'hist'}, defining which
type of plot.}

\item{qline}{if \code{type = 'hist'}, then \code{qline = TRUE} will
plot vertical lines at the 25/50/75th percentiles.}

\item{...}{further arguments passed to plotting functions.}
}
\value{
Plots to device.
}
\description{
Scatterplot or histogram faceted by groups.
}
\details{
Useful for disaggregating by some grouping variable, either as a
     bivariate scatterplot or univariate histogram.
}
\examples{
data(est)
x  <- calc_biomass(est)
fg <- summary_fg(x, eachplot=TRUE)
fg$logmass <- log10(fg$mass+1)
plot_facet(fg, 'cover', 'mass', by='fg', type='plot',
           xlab='Plot cover (\%)',
           ylab=bquote('Biomass (kg' ~ha^-1*')'))
plot_facet(fg, 'logmass', by='fg', type='hist', breaks=33,
           xlab='Biomass', ylab='Frequency')

}
