% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map of estimation plots}
\usage{
plot_map(x, xvar, log = TRUE, col, xlab, ylab, ...)
}
\arguments{
\item{x}{\code{'summary_plot'} object from function of the same
name.}

\item{xvar}{name of variable in \code{x} for which colors are
rendered.}

\item{log}{logical, use \code{'log10(x+1'} transformation (default)
or not.}

\item{col}{a vector of colors.  If missing (default) then colors
follow \code{\link[viridis]{inferno}}.}

\item{xlab, ylab}{optional axis labels.}

\item{...}{further arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
Plots to device.
}
\description{
Map geographic plot locations, colored by value.
}
\details{
Color values are relativized for each graphical image; plotting
     absolute values (when these differ among several plots) may
     require anually feeding a color vector in the proper order.
}
\examples{
data(est)
x <- calc_biomass(est)
s <- summary_plot(x)
plot_map(s, total_mn, main='Total biomass')

}
