% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fg.R
\name{summary_fg}
\alias{summary_fg}
\title{Summary by functional group}
\usage{
summary_fg(x, eachplot = FALSE, ...)
}
\arguments{
\item{x}{\code{"grlyr"} object from \code{calc_biomass()}.}

\item{eachplot}{default \code{FALSE} returns values as the mean
across all plots in \code{x}; \code{TRUE} returns values as
the mean of each functional group within each plot (and may
exclude functional groups unobserved on a given plot).}

\item{...}{further arguments (currently ignored).}
}
\value{
A data.frame with estimates for each functional group 'fg',
     including 'mass', 'masssd', 'c', 'csd', 'n', 'nsd' (all in kg
     per ha), as well as 'vol' and 'volsd' (cubic meters per ha),
     and 'cover' and 'coversd' (percentage points per plot).
}
\description{
Summary of ground layer observations, disaggregated by functional
     groups.
}
\details{
Gives basic descriptive summaries, disaggregated by functional
     groups.
}
\examples{
data(est)
x <- calc_biomass(est)
summary_fg(x)
summary_fg(x, eachplot=TRUE)

}
